Backend/LLM Backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv 
groq
cerebras_cloud_sdk